+++ client/tests/unit/utils/plasmid-map-test.js
@@ -1,0 +1,10 @@
+import plasmidMap from '../../../utils/plasmid-map';
+import { module, test } from 'qunit';
+
+module('plasmidMap');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  var result = plasmidMap();
+  assert.ok(result);
+});
